data_warehouse_api_automation/get-weather-api-function/requirements.txt
google-cloud-secret-manager==2.4.0
google-cloud-core==1.3.0
google-cloud-storage==1.25.0